# apps/backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the backend API for QuokkaQ, rewritten in Go.",
        "title": "QuokkaQ Go Backend API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:3001",
    "basePath": "/",
    "paths": {
...
        "/subscription-plans": {
            "get": {
                "description": "Returns all active subscription plans",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subscriptions"
                ],
                "summary": "Get Available Subscription Plans",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.SubscriptionPlan"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/subscriptions/checkout": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a checkout session for subscription upgrade",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subscriptions"
                ],
                "summary": "Create Checkout Session",
                "parameters": [
                    {
                        "description": "Checkout Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.CreateCheckoutRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.CreateCheckoutResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "501": {
                        "description": "Billing checkout not configured",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/subscriptions/me": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns subscription for the authenticated user's company",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subscriptions"
                ],
                "summary": "Get Current User's Subscription",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Subscription"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "No subscription found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/subscriptions/{id}/cancel": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Cancels the subscription at the end of billing period",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subscriptions"
                ],
                "summary": "Cancel Subscription",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Subscription ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Subscription"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.CreateCheckoutRequest": {
            "type": "object",
            "properties": {
                "planCode": {
                    "type": "string"
                }
            }
        },
        "handlers.CreateCheckoutResponse": {
            "type": "object",
            "properties": {
                "checkoutUrl": {
                    "type": "string"
                },
                "sessionId": {
                    "type": "string"
                }
            }
        },
...
        "models.Subscription": {
            "type": "object",
            "properties": {
                "cancelAtPeriodEnd": {
                    "type": "boolean"
                },
                "companyId": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "currentPeriodEnd": {
                    "type": "string"
                },
                "currentPeriodStart": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "metadata": {
                    "description": "additional data",
                    "type": "object"
                },
                "plan": {
                    "description": "Relations",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.SubscriptionPlan"
                        }
                    ]
                },
                "planId": {
                    "type": "string"
                },
                "status": {
                    "description": "\"trial\", \"active\", \"past_due\", \"canceled\", \"paused\"",
                    "type": "string"
                },
                "stripeSubscriptionId": {
                    "description": "Stripe sub_… after Checkout completes",
                    "type": "string"
                },
                "trialEnd": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "models.SubscriptionPlan": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "unique plan code like \"starter\", \"professional\"",
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "currency": {
                    "description": "\"RUB\", \"USD\"",
                    "type": "string"
                },
                "features": {
                    "description": "feature flags",
                    "type": "object"
                },
                "id": {
                    "type": "string"
                },
                "interval": {
                    "description": "\"month\", \"year\"",
                    "type": "string"
                },
                "isActive": {
                    "type": "boolean"
                },
                "limits": {
                    "description": "quota limits",
                    "type": "object"
                },
                "name": {
                    "description": "\"Starter\", \"Professional\", \"Enterprise\"",
                    "type": "string"
                },
                "price": {
                    "description": "price in minor units (cents/kopeks)",
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}